Refactor titles (#29)
* refactor titles

* remove useless override

* remove useless NxtSettingActivity
