+++ app/src/main/java/com/pileproject/drive/title/TitleActivityBase.java
@@ -16,31 +16,57 @@
 
 package com.pileproject.drive.title;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
+import android.view.View;
 
 import com.pileproject.drive.R;
 
 /**
- * Title Screen
- * Users can move to Programming Screen or Device Selection Screen.
- *
- * @author <a href="mailto:tatsuyaw0c@gmail.com">Tatsuya Iwanari</a>
- * @version 1.0 4-June-2013
+ * The base class of title pages.
+ * Users can move to programming page or preference page.
  */
 public abstract class TitleActivityBase extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_title);
+
+        // button to move to programming page
+        findViewById(R.id.title_startButton).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                startActivity(getIntentToProgrammingPage());
+            }
+        });
+
+        // button to move to preference page
+        findViewById(R.id.title_settingButton).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                startActivity(getIntentToSettingPage());
+            }
+        });
+
+        setupToolbar();
     }
 
     protected void setupToolbar() {
         Toolbar toolbar = (Toolbar) findViewById(R.id.title_toolbar);
-
         toolbar.setLogo(R.drawable.icon_launcher);
         toolbar.setTitle(R.string.app_name);
-
         setSupportActionBar(toolbar);
     }
+
+    /**
+     * A method to get an intent to programming page.
+     */
+    protected abstract Intent getIntentToProgrammingPage();
+
+    /**
+     * A method to get an intent to preference page.
+     */
+    protected abstract Intent getIntentToSettingPage();
 }
