+++ app/src/nxt/java/com/pileproject/drive/title/NxtTitleActivity.java
@@ -17,36 +17,21 @@
 package com.pileproject.drive.title;
 
 import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
 
-import com.pileproject.drive.R;
 import com.pileproject.drive.programming.visual.activity.NxtProgrammingActivity;
-import com.pileproject.drive.setting.NxtSettingActivity;
+import com.pileproject.drive.setting.SettingActivity;
 
+/**
+ * The title page of NxtDrive.
+ */
 public class NxtTitleActivity extends TitleActivityBase {
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_title);
-        // Button to move NxtProgrammingActivity
-        findViewById(R.id.title_startButton).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getApplicationContext(), NxtProgrammingActivity.class);
-                startActivity(intent);
-            }
-        });
+    protected Intent getIntentToProgrammingPage() {
+        return new Intent(getApplicationContext(), NxtProgrammingActivity.class);
+    }
 
-        // Button to move DeviceSelectActivity
-        findViewById(R.id.title_settingButton).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getApplicationContext(), NxtSettingActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        setupToolbar();
+    @Override
+    protected Intent getIntentToSettingPage() {
+        return new Intent(getApplicationContext(), SettingActivity.class);
     }
 }
